+++ src/operators/opIntEq.h
@@ -18,7 +18,7 @@ struct OpIntEq : public BoolView {
     void evaluate() final;
     void startTriggering() final;
     void stopTriggering() final;
-    void updateViolationDescription(u_int64_t parentViolation,
+    void updateViolationDescription(UInt parentViolation,
                                     ViolationDescription&) final;
     ExprRef<BoolView> deepCopySelfForUnroll(
         const AnyIterRef& iterator) const final;
