+++ src/operators/opSetSize.h
@@ -13,7 +13,7 @@ struct OpSetSize : public IntView {
     void evaluate() final;
     void startTriggering() final;
     void stopTriggering() final;
-    void updateViolationDescription(u_int64_t parentViolation,
+    void updateViolationDescription(UInt parentViolation,
                                     ViolationDescription&) final;
     ExprRef<IntView> deepCopySelfForUnroll(
         const AnyIterRef& iterator) const final;
